More work on message bases

+++ jamlib-1.4.7/include/jamlib/jam.h
@@ -49,7 +49,7 @@
 #if !(((defined(__FreeBSD__) && __FreeBSD_version >= 440000)) || defined(NeXTBSD))
 typedef unsigned short ushort;   /* must be 16 bits wide */
 #endif
-typedef unsigned long  ulong;    /* must be 32 bits wide */
+typedef uint32_t  ulong;    /* must be 32 bits wide */
 #endif
 typedef unsigned char  uchar;    /* must be  8 bits wide */
